Move glossary definitions to data file
Moves all glossary definitions to a glossary.yml file in the data
directory, so that definitions can be pulled into the site pages
by referencing the glossary data file. This will allow for more
easy access to the glossary definitions. This also changes the
glossary page to pull in the definitions from the glossary data file.

Signed-off-by: Shannyn Telander <telander@bitwise.io>